Moving parsePositiveQueryInt to a React free file to give better dependency isolation
