+++ src/lib/paging.ts
@@ -18,17 +18,6 @@ export interface SwrProps {
 type QueryValue = number | string | undefined;
 const QueryParamOrder = ["pageSize", "cursor", "sort", "suppliedId", "name"];
 
-export function parsePositiveQueryInt(
-  value: string | undefined,
-  defaultValue: number
-): number {
-  const trimmed = value?.trim();
-  if (trimmed == null || !/^\d+$/.test(trimmed)) return defaultValue;
-
-  const parsed = Number.parseInt(trimmed, 10);
-  return parsed > 0 ? parsed : defaultValue;
-}
-
 export function buildQuery(
   path: string,
   params: Record<string, QueryValue>
