+++ src/lib/query-params.ts
@@ -1,0 +1,10 @@
+export function parsePositiveQueryInt(
+  value: string | undefined,
+  defaultValue: number
+): number {
+  const trimmed = value?.trim();
+  if (trimmed == null || !/^\d+$/.test(trimmed)) return defaultValue;
+
+  const parsed = Number.parseInt(trimmed, 10);
+  return parsed > 0 ? parsed : defaultValue;
+}
