+++ src/pages/api/file-collections.ts
@@ -20,7 +20,7 @@ import {
   toErrorRes,
 } from "../../lib/api";
 import { getFileCollectionsApi } from "../../lib/file-collections";
-import { parsePositiveQueryInt } from "../../lib/paging";
+import { parsePositiveQueryInt } from "../../lib/query-params";
 import { getClientFromSession, makeCall } from "../../lib/vertex-api";
 import withSession, { NextIronRequest } from "../../lib/with-session";
