+++ src/pages/api/part-revisions.ts
@@ -14,7 +14,7 @@ import {
   ServerError,
   toErrorRes,
 } from "../../lib/api";
-import { parsePositiveQueryInt } from "../../lib/paging";
+import { parsePositiveQueryInt } from "../../lib/query-params";
 import { getClientFromSession } from "../../lib/vertex-api";
 import withSession, { NextIronRequest } from "../../lib/with-session";
